Simplify test data for temporary edges
