+++ street/src/test-fixtures/java/org/opentripplanner/street/model/edge/SimpleConcreteEdge.java
@@ -4,6 +4,7 @@ import org.opentripplanner.core.model.i18n.I18NString;
 import org.opentripplanner.street.geometry.SphericalDistanceLibrary;
 import org.opentripplanner.street.model.vertex.Vertex;
 import org.opentripplanner.street.search.TraverseMode;
+import org.opentripplanner.street.search.request.StreetSearchRequest;
 import org.opentripplanner.street.search.state.State;
 import org.opentripplanner.street.search.state.StateEditor;
 
@@ -24,7 +25,7 @@ public class SimpleConcreteEdge extends Edge {
   public State[] traverse(State s0) {
     double d = getDistanceMeters();
     TraverseMode mode = s0.currentMode();
-    int t = (int) ((1000.0 * d) / TemporaryConcreteEdge.getSpeed(mode, s0.getRequest(), false));
+    int t = (int) ((1000.0 * d) / getSpeed(mode, s0.getRequest(), false));
     StateEditor s1 = s0.edit(this);
     s1.incrementTimeInMilliseconds(t);
     s1.incrementWeight(d);
@@ -42,5 +43,18 @@ public class SimpleConcreteEdge extends Edge {
       getFromVertex().getCoordinate(),
       getToVertex().getCoordinate()
     );
+  }
+
+  /**
+   * The road speed for a specific traverse mode.
+   *
+   */
+  public static double getSpeed(TraverseMode mode, StreetSearchRequest req, boolean walkingBike) {
+    return switch (mode) {
+      case WALK -> walkingBike ? req.bike().walking().speed() : req.walk().speed();
+      case BICYCLE -> req.bike().speed();
+      case SCOOTER -> req.scooter().speed();
+      default -> throw new IllegalArgumentException("getSpeed(): Invalid mode " + mode);
+    };
   }
 }
